+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Added `nstd_core_cstr_as_slice`.
 - Moved the `cstr` module out of `str`.
 ### `nstd.env`
+- `nstd_env_vars` now returns a vector of `NSTDString`.
 - `nstd_env_remove_var` now uses `NSTDChar`.
 - `nstd_env_get_var` now uses `NSTD[Char|String]`.
 - `nstd_env_set_var` now uses `NSTDChar`.